+++ src/main.c
@@ -1,5 +1,6 @@
 #include "../include/utils.h"
 #include "../include/sort_data.h"
+#include "../include/sort_data_thread.h"
 #include "../include/find_average_salary.h"
 #include <stdlib.h>
 
@@ -17,32 +18,19 @@ int main(int argc, char* argv[]) {
         return CANNOT_GET_INFO_FROM_FILE;
     }
 
+   
     main_list_t * head = NULL;
-    head = initialise_main_list(workers[0]);
-    if (head == NULL) {
-        return CANNOT_BUILD_LIST_STRUCTURE;
-    }
-    for(size_t i = 1; i < size; i++) {
+        
+    for(size_t i = 0; i < size; i++) {
         head = add_elem_to_main_structure(workers[i], head);
     }
-
-    if (head == NULL) {
-        return ADDING_ERROR;
-    }
-
-    if (print_info(workers, &size)) {
-        return CANNOT_PRINT_INFO_FROM_ARRAY;
-    }
-    if (print_position_structure(head)) {
-        return CANNOT_PRINT_INFO_FROM_LIST_STRUCTURE;
-    }
-
-    if (find_average_salary(head)) {
-        return ERROR_IN_BUILDING_AVERAGE_SALARY_MODEL;
-    }
-    if (clear_position_structure(head)) {
-        return ERROR_IN_DELETING_LIST_STRUCTURE;
-    }
+    
+    if (strcmp(argv[2], "1") == 0)
+    parallel_prog(head);
+    if (strcmp(argv[2], "1") == 1)
+    find_average_salary(head);
+    sleep(1);
+   
 
     free(workers);  // по логике она наверно в памяти должна оставаться
     return NO_ERROR;
