+++ test/CMakeLists.txt
@@ -1,4 +1,4 @@
-set(CMAKE_CXX_FLAGS "-fprofile-arcs -ftest-coverage --coverage")
+set(CMAKE_CXX_FLAGS "-fprofile-arcs -ftest-coverage --coverage -pthread")
 set(CMAKE_CXX_STANDARD 17)
 include_directories("${PROJECT_SOURCE_DIR}/include")
